+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.4.4.0 (unreleased)
 --------------------
 - Limit all operations to simple and configurable products
+- Support discount codes
 
 1.4.3.0
 -------